feat: Add ignored columns reporting
This commit introduces a new feature that provides reporting
of ignored columns when the corresponding option is enabled.

A new option, `report_ignored_columns`, has been added to the
`Sheetah::Template` initialization. It works in conjunction
with the existing `ignore_unspecified_columns` option, which
must also be enabled for the reporting to take effect.